Rizky

+++ controllers/recipe.js
@@ -1,4 +1,4 @@
-const { Recipe } = require('../models/')
+const { Recipe, UserRecipe, Tag, RecipeTag } = require('../models/')
 const { gql } = require('apollo-server')
 
 
@@ -30,8 +30,8 @@ const typeDefs = gql`
       recipes: [Recipe]
     }
     extend type Mutation {
-      addRecipe(recipe: NewRecipe): Recipe
-      editRecipe(id: Int! recipe: NewRecipe): Recipe
+      addRecipe(recipe: NewRecipe, tags: [String!]): Recipe
+      editRecipe(id: Int! recipe: NewRecipe, tags: [String!]): Recipe
       deleteRecipe(id: Int!): Response
     }`
 
@@ -40,7 +40,7 @@ const resolvers = {
   Query: {
     recipe: async (_, args) => {
       try {
-        const data = await Recipe.findOne({ where: { id: args.id } })
+        const data = await Recipe.findOne({ where: { UserId: args.id }, include: { model: Tag } })
         return data
       } catch (error) {
         console.log(error)
@@ -48,7 +48,7 @@ const resolvers = {
     },
     recipes: async () => {
       try {
-        const data = await Recipe.findAll()
+        const data = await Recipe.findAll({ include: { model: Tag } })
         return data
       } catch (error) {
         console.log(error)
@@ -56,26 +56,78 @@ const resolvers = {
     }
   },
   Mutation: {
-    addRecipe: async (_, args) => {
+    addRecipe: async (_, args, context) => {
       try {
-        const data = await Recipe.create(args.recipe)
-        return data
+        if (!context.user) throw new AuthenticationError("Please login first");
+        const { id } = context.user;
+        const data = await Recipe.create(args.recipe);
+        const dataUserRecipe = {
+          UserId: id,
+          RecipeId: data.id,
+          favorites: false,
+          plan: [],
+          creation: true
+        };
+        const userRecipe = await UserRecipe.create(dataUserRecipe);
+        const { tags } = args;
+        for (i in tags) {
+          let newTag = await Tag.findOrCreate({
+            where: {
+              name: tags[i].trim().toLowerCase()
+            }
+          });
+          let newPayload = {
+            RecipeId: data.id,
+            TagId: newTag[0].id
+          };
+          await RecipeTag.create(newPayload);
+        }
+        const result = await Recipe.findByPk(data.id, {
+          include: Tag
+        })
+        console.log(result)
+        return result;
       } catch (error) {
         console.log(error)
       }
     },
-    editRecipe: async (_, args) => {
+    editRecipe: async (_, args, context) => {
       try {
-        const data = await Recipe.update(args.recipe, { where: { id: args.id }, returning: true })
-        return data[1][0].dataValues
+        if (!context.user) throw new AuthenticationError("Please login first");
+        const data = await Recipe.update(args.recipe, {
+          where: { id: args.id },
+          returning: true
+        });
+        const { tags } = args;
+        await RecipeTag.destroy({ where: { RecipeId: data[1][0].id } })
+        for (i in tags) {
+          let newTag = await Tag.findOrCreate({
+            where: {
+              name: tags[i].trim().toLowerCase()
+            }
+          });
+          let newPayload = {
+            RecipeId: data[1][0].id,
+            TagId: newTag[0].id
+          };
+          await RecipeTag.findOrCreate({
+            where: newPayload
+          });
+        }
+        const result = await Recipe.findByPk(data[1][0].id, {
+          include: Tag
+        })
+        return result;
       } catch (error) {
         console.log(error)
       }
     },
-    deleteRecipe: async (_, args) => {
+    deleteRecipe: async (_, args, context) => {
       try {
-        await Recipe.destroy({ where: { id: args.id } })
-        return { message: 'Recipe has been deleted' }
+        if (!context.user) throw new AuthenticationError("Please login first");
+        await Recipe.destroy({ where: { id: args.id } });
+        await RecipeTag.destroy({ where: { RecipeId: args.id } })
+        return { message: "Recipe has been deleted" };
       } catch (error) {
         console.log(error)
       }
